+++ examples/statistics/ANCOVA_rutherford.py
@@ -24,7 +24,7 @@ cov = Var([9, 5, 6, 4, 6, 8, 3, 5,
            8, 5, 6, 5, 3, 6, 4, 6,
            5, 8, 3, 4, 6, 9, 4, 5], name='cov')
 
-A = Factor([1, 2, 3], rep=8, name='A' )
+A = Factor([1, 2, 3], repeat=8, name='A' )
 
 print test.anova(Y, A + cov)
 print test.anova(Y, cov * A)